fix: refactor andfix add script modal
